try using .exit instead of .incomplete
We could forget to remote the `*done` file on a bad run (or maybe we hit
a bug). And there could be other race conditions.

`*done.exit` is clear, and it lets other tasks depend on `*done` without
any possible race condition.

This does *not* solve the problem of Grid jobs which are killed
directly.