Fix SLURM support for FALCON
Had to come up with a simple way to fix the "Missing interpreter" error SLURM was giving with FALCON 0.3 that wouldn't break other schedulers. This should make FALCON stable for any future releases, and is compatible with Python 3.0+ as well, so no need to update code if/when Python version is updated.

Update run.py

Add SCANCEL support

Add support for SLURM jobs to automatically cancel failed jobs in event of FALCON failure.

fix indentation and loss of break

Update run.py

closes FALCON#181