Reduce typical job_uid
This is a favor to @dgordon562. His SGE names are truncated when
printed. We simply drop some leading zeroes and use hexadecimal.
If the number of jobs gets huge, we just lose the leading zeroes
and fixed width, no problem. But if you have > 64k jobs, you have
bigger problems.