Off-by-1
Pretty harmless error in the recent max_cov_aln code.
Actually, the error was in the older code.
We were counting the initial seed among the min/max number of reads,
but the seed can occur among the reads for consensus too. So we should
not count the initial seq (the "seed") when we compute the number
of sequences used for consensus. But the error was already made,
so now we need to retain the meaning of existing config-files.