move ext_falcon.so to root
I wanted to hide it under the falcon_kit package namespace, but
that turns out to be problematic.

If you run python from the root directory, then no matter how you installed, you will be using the package from `.`,
since python prepends `''` at start-up. But the extension module is not there. So it needs to be loaded
via absolute path. Not really a big deal. `ext_falcon.so` is uniquish enough.