Avoid problem in unzip
This is for a different problem, discussed between Jason and Dario. I am not certain of the problem yet. From Jason's email:

For FALCON-unzip, you can comment out the following lines in graph_to_contig.py to avoid the seg fault. The FALCON-Unzip will run. Your a_ctg.fa will have some duplicates but it will not affects downstream in general. The detail of a_tig is not used in Unzip process as it replaces it with phased haplotigs.

Ok, the root cause is that there is a very long bubble.  The alternative path is > 500kb.  The internal aligner used is not designed for handling such long sequences. In a halpoid genome, this is not supposed to happened. However, in Cab-Suv genome, we might generate big bubble in the graph level.  One proposed fix is to cap the size of sequence sent to the internal aligner now.