Fixed the gfa_from_assembly method to omit deduplicated a_ctg.fa
Moved the part of code which loads tiling paths into the gfa_graph into a separate function.

Modifying a wrapper method for the GFA writer. This commit only moves file path definitions to the top.

This commit removes the manual loading of tiling paths in order to test the writing procedure. This part was extracted into a separate method in gen_gfa_v1.py script, which can now be reused. All tests are now passing.

autopep8