Wrap all parallel executables to be "chunkable"
A wrapper should take as input a JSON list of records (units of work),
and it should execute each record in series. That way, when pbsmrtpipe
"chunks", the input still works. For us, the input will usually
be a list of one element.

This is implemented as an extra task in the pypeflow graph before each
wrapped task. That is because we want to use bash scripts for task, and
we cannot send an arbitrary number of filenames to a bash command-line.
(pbsmrtpipe uses "resolved tool contracts" to pass a list of filenames,
but that solution would couple us tightly to pbsmrtpipe.)
See `gen_parallel_tasks` for more details.

(Snakemake is currently not supported.)

Hopefully, we can simplify this more eventually. But the main
goal is to make our tasks 1:1 with pbsmrtpipe in a simple way.